Change to template functions
Instead of function naming to avoid code duplication.
